Added working functionality to submit button for sign in page. Will open a new page if correct info is used
