+++ src/index.html
@@ -20,7 +20,7 @@ This will be the sign on screen
 
 	<img src="../assets/images/logoWithTextUnder.png" id="logoWithTextUnder">
 	<div class="jumbotron">
-		<form>
+		<form onSubmit="return false;">
 		  <div class="form-group">
 		    <label for="signInUsername">Username</label>
 		    <input type="Username" class="form-control" id="signInUsername" placeholder="Enter Username">
@@ -34,6 +34,8 @@ This will be the sign on screen
 	</div>
 
 	<script src="index.js"></script>
+
+	<script src="appHomePage.js"></script>
 	<!-- 
 	Required Javascript scripts for using Bootstrap and jQuery to build faster and more appealing applications
 	-->
